test_job_enrichment.py
"""
Test script for job enrichment feature
Tests database, Claude integration, and workflow
"""

import os
import sys
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

print("=" * 60)
print("Testing LeadOn CRM Job Enrichment Feature")
print("=" * 60)
print()

# Test 1: Database
print("Test 1: Database Setup")
print("-" * 60)
try:
    from database.db_manager import get_db_manager
    from database.models import Company, Contact, JobPosting
    
    db = get_db_manager()
    session = db.get_session()
    
    print("✅ Database initialized successfully")
    print(f"   Database location: database/leadon.db")
    
    # Test creating a company
    test_company, created = db.get_or_create_company(
        session,
        name="Test Company Inc",
        description="A test company for verification",
        industry="Technology",
        source="test"
    )
    
    if created:
        print(f"✅ Created test company: {test_company.name}")
    else:
        print(f"✅ Test company already exists: {test_company.name}")
    
    session.close()
    
except Exception as e:
    print(f"❌ Database test failed: {e}")
    sys.exit(1)

print()

# Test 2: Claude API
print("Test 2: Claude API")
print("-" * 60)
try:
    from ai_agent.intent_parser import IntentParser
    
    parser = IntentParser()
    print("✅ Claude client initialized")
    print(f"   Model: claude-3-haiku-20240307")
    
    # Test a simple query
    response = parser.client.messages.create(
        model="claude-3-haiku-20240307",
        max_tokens=50,
        messages=[{"role": "user", "content": "Say 'test successful' in one sentence"}]
    )
    
    print(f"✅ Claude API working")
    print(f"   Response: {response.content[0].text}")
    
except Exception as e:
    print(f"❌ Claude API test failed: {e}")
    sys.exit(1)

print()

# Test 3: Apollo API
print("Test 3: Apollo API")
print("-" * 60)
try:
    from scrapers.apollo_scraper import ApolloClient
    
    if not os.getenv("APOLLO_API_KEY"):
        print("⚠️  Apollo API key not set")
    else:
        client = ApolloClient()
        print("✅ Apollo client initialized")
        print(f"   API Key: {os.getenv('APOLLO_API_KEY')[:15]}...")
        
        # Test a simple search
        result = client.search_people(
            titles=["CEO"],
            per_page=1
        )
        
        print(f"✅ Apollo API working")
        print(f"   Total contacts available: {result.total_results:,}")
        
except Exception as e:
    print(f"❌ Apollo API test failed: {e}")

print()

# Test 4: Job Enrichment Service
print("Test 4: Job Enrichment Service")
print("-" * 60)
try:
    from services.job_enrichment_service import JobEnrichmentService
    from scrapers.apollo_scraper import ApolloClient
    from ai_agent.intent_parser import IntentParser
    from database.db_manager import get_db_manager
    
    apollo = ApolloClient()
    claude = IntentParser()
    db = get_db_manager()
    
    service = JobEnrichmentService(apollo, claude, db)
    print("✅ Job enrichment service initialized")
    
    # Test query generation
    queries = service.generate_job_search_queries(
        user_query="Find companies that need sales automation",
        product_description="Sales automation platform"
    )
    
    print(f"✅ Query generation working")
    print(f"   Generated {len(queries)} job search queries:")
    for q in queries[:3]:
        print(f"      - {q['query']} in {q['location']}")
    
except Exception as e:
    print(f"❌ Job enrichment service test failed: {e}")

print()

# Test 5: LinkedIn Scraper
print("Test 5: LinkedIn Scraper")
print("-" * 60)
try:
    from linkedin_scrape import scrape_first_n_jobs
    
    print("✅ LinkedIn scraper module loaded")
    print("   Note: Actual scraping not tested to avoid rate limits")
    print("   Functions available:")
    print("      - scrape_first_n_jobs()")
    print("      - fetch_job_detail()")
    
except Exception as e:
    print(f"❌ LinkedIn scraper test failed: {e}")

print()

# Summary
print("=" * 60)
print("Test Summary")
print("=" * 60)
print()
print("✅ All core components are working!")
print()
print("Next steps:")
print("1. Start the server: python crm_integration\\chat_api.py")
print("2. Open browser: http://localhost:8000")
print("3. Try a search with job enrichment enabled")
print()
print("Example query:")
print('   "Find companies I can sell my sales automation product to"')
print("   ✅ Check: Enrich with Job Postings")
print('   Product: "Sales automation platform for B2B companies"')
print()
print("=" * 60)

